Initial Commit
Was helping out a friend with a final project they were working on.
Decided best way to not lead them astray while guiding them along was to
make sure my header was on the level and that things would work out. So
I made a parrallel system while they wrote theirs. Mind has more doodads
and uses spinners and such while theirs was just text fields, but it was
fun playing with swing again. Been such a long time.

The layout looks like crap, but hey, it's just an instructional example,
not a production grade application.